cleanup, dev 1

+++ src/main/java/com/parzi/starwarsmod/jedirobes/powers/PowerJump.java
@@ -1,10 +1,6 @@
 package com.parzi.starwarsmod.jedirobes.powers;
 
-import com.parzi.starwarsmod.utils.PotionList;
-
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.potion.Potion;
-import net.minecraft.potion.PotionEffect;
 import net.minecraft.util.MathHelper;
 
 public class PowerJump extends Power
@@ -27,13 +23,13 @@ public class PowerJump extends Power
 		{
 			player.motionY = 0.41999998688697815D;
 
-			player.motionY += (double)((float)(this.currentLevel + 1) * 0.11F);
+			player.motionY += (this.currentLevel + 1) * 0.11F;
 
 			if (player.isSprinting())
 			{
 				float f = player.rotationYaw * 0.017453292F;
-				player.motionX -= (double)(MathHelper.sin(f) * 0.2F);
-				player.motionZ += (double)(MathHelper.cos(f) * 0.2F);
+				player.motionX -= MathHelper.sin(f) * 0.2F;
+				player.motionZ += MathHelper.cos(f) * 0.2F;
 			}
 
 			player.isAirBorne = true;
